+++ lib/ex_daytona/api/health.ex
@@ -19,14 +19,14 @@ defmodule ExDaytona.Api.Health do
 
   ### Returns
 
-  - `{:ok, ExDaytona.Model.HealthControllerCheck200Response.t}` on success
+  - `{:ok, ExDaytona.Model.Check200Response.t}` on success
   - `{:error, Tesla.Env.t}` on failure
   """
-  @spec health_controller_check(Tesla.Env.client(), keyword()) ::
-          {:ok, ExDaytona.Model.HealthControllerCheck200Response.t()}
-          | {:ok, ExDaytona.Model.HealthControllerCheck503Response.t()}
+  @spec check(Tesla.Env.client(), keyword()) ::
+          {:ok, ExDaytona.Model.Check200Response.t()}
+          | {:ok, ExDaytona.Model.Check503Response.t()}
           | {:error, Tesla.Env.t()}
-  def health_controller_check(connection, _opts \\ []) do
+  def check(connection, _opts \\ []) do
     request =
       %{}
       |> method(:get)
@@ -36,8 +36,8 @@ defmodule ExDaytona.Api.Health do
     connection
     |> Connection.request(request)
     |> evaluate_response([
-      {200, ExDaytona.Model.HealthControllerCheck200Response},
-      {503, ExDaytona.Model.HealthControllerCheck503Response}
+      {200, ExDaytona.Model.Check200Response},
+      {503, ExDaytona.Model.Check503Response}
     ])
   end
 
@@ -53,8 +53,8 @@ defmodule ExDaytona.Api.Health do
   - `{:ok, nil}` on success
   - `{:error, Tesla.Env.t}` on failure
   """
-  @spec health_controller_live(Tesla.Env.client(), keyword()) :: {:ok, nil} | {:error, Tesla.Env.t()}
-  def health_controller_live(connection, _opts \\ []) do
+  @spec live(Tesla.Env.client(), keyword()) :: {:ok, nil} | {:error, Tesla.Env.t()}
+  def live(connection, _opts \\ []) do
     request =
       %{}
       |> method(:get)